backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.4
Faker==22.0.0
